+++ src/docs/docs.module.js
@@ -1,0 +1,11 @@
+'use strict';
+
+require('jsonforms/dist/jsonforms.css');
+require('jsonforms/dist/jsonforms-material.css');
+
+require('jsonforms');
+require('jsonforms/material/jsonforms-material');
+
+require('./uischema/uischema.controller');
+
+angular.module('docs', []);
